refactor: use virtual file to wrap components
