+++ test/integration/fixtures/components/Counter.vue
