+++ test/integration/fixtures/components/CounterA.vue
