+++ test/integration/fixtures/components/CounterB.vue
