+++ test/integration/fixtures/components/Inner.vue
