+++ test/integration/fixtures/components/Outer.vue
@@ -1,10 +1,10 @@
 <script setup>
-import Inner from './Inner.island.vue'
+import Inner from './Inner.vue'
 </script>
 
 <template>
   <div class="outer">
     <span>Outer</span>
-    <Inner />
+    <Inner v-client:load />
   </div>
 </template>
