+++ test/integration/fixtures/components/StyledCounter.vue
