+++ test/integration/fixtures/components/WithIslandProps.vue
@@ -1,9 +1,9 @@
 <script setup>
-import Counter from './Counter.island.vue'
+import Counter from './Counter.vue'
 </script>
 
 <template>
   <div>
-    <Counter :count="42" label="Click me" />
+    <Counter v-client:load :count="42" label="Click me" />
   </div>
 </template>
